feat(cli): PNPM support #15055 (#15198) + small cmd refactoring
